Ah, the early stages of awkward youth!
The unit tests which are here, such as they are, pass.
That just means that the most basic functionality is beginning
to come together, but we're still so far pre-alpha that the
architecture has yet to be solidified.  There's a ton of work
left to do polishing off the rough edges of the ideas themselves,
prior to even considering their implementation.

Which is to say, "don't do anything with this yet", and "please
reserve judgment, we know this is a goofy pre-toddler of a
project."