Removing prototypes from methods
...because prototypes apply to functions, not methods.  Duh.
And they were starting to run afoul of the Class::Method::Modifiers
being used for safety checking some operations, which means the
nice (but pointless) visual hint was starting o caseu actual
warnings, and thus problems, so they had to go.  Which means more
emphasis will need to be placed on good docs and arg checks.