IO::All port for file management, code organization.
IO::All is more thorough than I'm going to be for file management, and
looks to be reasonably well supported, so let's go with that.

Adds the meta attribute to documents so we can start doing safety checks
like storing the timestamp of the last time it was modified, and making
appropriate amounts of noise if it's changed before we try and change
it, etc.

Expanded POD, renamed a few things, moved a few other things around,
generally trying to be tidy.  Still a loooong way to go.