POD driven refactoring of indexing methods
Expands somewhat in the "INDEXING" POD section and examples.

Delegates inversion of indexes back to strings over to its own method;
the abstraction (and thus extensibility in descendants) is worth the
overhead of additional arg checking and the like.

Renamed lookup_by_index into "search", which is the only differentiation
from "search_index", though they still go to the same place.

Tidied up actual index searching to retain both filenames and complete
paths, and be able to do path list returns and work in both scalar and
list contexts.