Lots more POD, not as much refactoring.
Added several additional POD entries, removed the standardize_index_name
method (this already has to match field names in the document, and we'll
just use those "as is"; we'll see if this proves foolhardy and needs
reinstatement later).

Makes suffix removal during key unindexing conditional on actually
having that suffix.

Only one exposed method left in Fileplex for documentation; internal
docs will be added to utilities, and then it's on to unit tests (and POD
for other classes).